[BUILD] Rename feature,update -> eclipse.feature,update
The new project names make clear, that the projects
'feature' and 'update' are related to the eclipse plugin.
Everybody that is familiar with the concept of update-sites
and features is directly able to identify the corresponding projects.
Furthermore the name feature/update without the eclipse context
is to generic in the context of software development and causes confusion.